correction de la sérialisation json avec une deep sérialisation

+++ src/main/resources/db/migration/V5__createVolatileStorageTable.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS volatile_storage(
     id SERIAL PRIMARY KEY,
     reference VARCHAR(255) NOT NULL,
-    available int NOT NULL,
-    frequency int,
-    latency int,
+    available DOUBLE PRECISION NOT NULL,
+    frequency DOUBLE PRECISION,
+    latency DOUBLE PRECISION,
     id_machine int NOT NULL,
     CONSTRAINT fk_vol_storage_machine FOREIGN KEY (id_machine) REFERENCES machine_entity (id)
 );
